Move class defn from ds-test.cpp to ds.h
Rationale: I want the spy subclass DiamondSquareSpy to be part of the
production code, not the test code.  This despite the fact that this
class exists solely for testing purposes: new functionality added to
the parent class must be mirrored in changes to the subclass.

Since the changes occur together, I put the classes in the same file.